add Question component to reuse in SingleQuestion and DoubleQuestion pages
