+++ components/Question/styles.module.css
@@ -1,0 +1,9 @@
+.container {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background: var(--light-blue-gradient);
+  width: 100%;
+  height: 100vh;
+}
